many: split out toml config package into separate sub-package
Also refactor the config handling so that the details a particular config is encapsulated outside of the toml configurator package. This ensures total generic operation with any config struct so long as it works with the upsteam toml package.
Doing this requires declaring the package's upstream location in the go.mod so that the tomlconfigurator package can be imported from cmd/server/main.go